refacto: update Proxy test and rename it in UserTest to clarify2
